Moved Main Tab View to view group and make a dedicated file
