+++ KiddyEatsApp/KiddyEatsApp.swift
@@ -11,45 +11,9 @@ import SwiftUI
 struct KiddyEatsApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            MainTabView()
+                .environment(ExploreViewModel())
         }
     }
 }
 
-struct ContentView: View {
-    var body: some View {
-        MainTabView()
-            .environment(ExploreViewModel())
-    }
-}
-
-struct MainTabView: View {
-    var body: some View {
-        TabView {
-            NavigationView {
-                ExploreView()
-            }
-            .tabItem {
-                Label("Explore", systemImage: "magnifyingglass.circle")
-            }
-            
-            NavigationView {
-                Text("Hello")
-            }
-            .tabItem {
-                Label("Collections", systemImage: "heart.circle")
-            }
-            
-            NavigationView {
-                Text("Hello")
-            }
-            .tabItem {
-                Label("Profile", systemImage: "person.circle")
-            }
-        }
-    }
-}
-
-#Preview {
-    ContentView()
-}
