+++ KiddyEatsApp/View/Main/MainTabView.swift
@@ -1,0 +1,39 @@
+//
+//  MainTabView.swift
+//  KiddyEatsApp
+//
+//  Created by Hans Arthur Cupiterson on 19/08/24.
+//
+
+import SwiftUI
+
+struct MainTabView: View {
+    var body: some View {
+        TabView {
+            NavigationView {
+                ExploreView()
+            }
+            .tabItem {
+                Label("Explore", systemImage: "magnifyingglass.circle")
+            }
+            
+            NavigationView {
+                Text("Hello")
+            }
+            .tabItem {
+                Label("Collections", systemImage: "heart.circle")
+            }
+            
+            NavigationView {
+                Text("Hello")
+            }
+            .tabItem {
+                Label("Profile", systemImage: "person.circle")
+            }
+        }
+    }
+}
+
+#Preview {
+    MainTabView()
+}
